+++ src/CombPermUtils.cpp
@@ -191,11 +191,11 @@ double MultisetPermRowNum(int n, int r, std::vector<int> &myReps) {
 
 // This algorithm is nearly identical to the
 // one found in the standard algorithm library
-void nextFullPerm(uint16_t *myArray, unsigned long int &n1,
+void nextFullPerm(int *myArray, unsigned long int &n1,
                   unsigned long int &n2) {
     
     unsigned long int p1 = n2, p2 = n1;
-    uint16_t temp;
+    int temp;
     
     while (myArray[p1 + 1] <= myArray[p1])
         --p1;
@@ -227,11 +227,11 @@ void nextFullPerm(uint16_t *myArray, unsigned long int &n1,
 // and swap them. We can then proceed to the next perm.
 // We can do this because the standard algo would end
 // up performing two unnecessary reversings.
-void nextPartialPerm(uint16_t *myArray, unsigned long int &r, 
+void nextPartialPerm(int *myArray, unsigned long int &r, 
                      unsigned long int &r1, unsigned long int &n,
                      unsigned long int &lastElem) {
     
-    uint16_t temp;
+    int temp;
     unsigned long int p1 = r1;
     
     while (p1 < n && myArray[r1] >= myArray[p1])
